Move volunteer submissions into proper place
