+++ src/application/VolunteerSubmissionsService.ts
@@ -1,0 +1,27 @@
+import FormVolunteerSubmission, { VolunteerFormStatus } from '@infrastructure/postgres/FormVolunteerSubmission';
+import { Repository } from 'typeorm';
+
+export enum FormStatus {
+    IN_PROGRESS = 'inProgress',
+    REJECTED = 'rejected',
+    ACCEPTED = 'accepted',
+}
+
+export interface ChangeStatusForVolunterFormParams {
+    status: VolunteerFormStatus;
+    userId: number;
+}
+
+export class VolunteerSubmissionsService {
+    constructor(private volunteerSubmissionRepository: Repository<FormVolunteerSubmission>) {}
+
+    public async changeStatusForVolunteerForm(changeStatusParams: ChangeStatusForVolunterFormParams): Promise<void> {
+        await this.volunteerSubmissionRepository
+            .createQueryBuilder()
+            .update()
+            .set({ status: changeStatusParams.status })
+            .where('userId = :id', { id: changeStatusParams.userId })
+            .execute();
+        return;
+    }
+}
