+++ src/presentation/rest/VolunteerSubmissionsController.ts
@@ -1,0 +1,20 @@
+import {
+    ChangeStatusForVolunterFormParams,
+    VolunteerSubmissionsService,
+} from '@application/VolunteerSubmissionsService';
+import { Body, Put, Route, Tags } from 'tsoa';
+import { Inject } from 'typescript-ioc';
+
+@Tags('Volunteer Submissions')
+@Route('volunteerSubmissions')
+export class AnimalSubmissionsController {
+    @Inject
+    private submissionService!: VolunteerSubmissionsService;
+
+    @Put('changeVolunterFormStatus')
+    public async changeFormStatusForVolunteer(
+        @Body() changeStatusParams: ChangeStatusForVolunterFormParams,
+    ): Promise<void> {
+        await this.submissionService.changeStatusForVolunteerForm(changeStatusParams);
+    }
+}
